+++ CHANGES.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- LoRA
-- transformer weights dtype
+- LoRA.
+- Transformer weights dtype.
+- User-requested max tokenizer length (overrides HF max model length).
 
 ### Added